# StrnadiAPI-openapi.yaml
openapi: 3.0.1
info:
  title: Strnadi API
  description: |
    HTTP API used by the Strnadi applications to manage user accounts, bird recordings,
    articles and other shared resources. The specification documents the behavior of the
    existing ASP.NET controllers so it can be imported into Swagger UI or other OpenAPI tooling.
  version: 1.0.0
servers:
  - url: https://api.strnadi.cz
    description: PROD API
  - url: https://devapi.strnadi.cz
    description: DEV API
security:
  - bearerAuth: []
...
paths:
  /auth/verify-jwt:
    get:
      tags: [Auth]
      summary: Verify that a JWT belongs to a verified account
      description: |
        Validates the bearer token supplied in the `Authorization` header and confirms that the
        underlying account has its email address verified.
      responses:
        '200':
          description: JWT is valid and the email has already been verified.
        '400':
          description: The bearer token is missing or the email could not be extracted.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: The JWT is valid but the email has not been verified yet.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/renew-jwt:
    get:
      tags: [Auth]
      summary: Renew a JWT
      description: Issues a fresh JWT for the currently authenticated user.
      responses:
        '200':
          description: Renewed JWT token.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/JwtToken'
        '400':
          description: The request did not provide a bearer token or the token was invalid.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: The user referenced by the JWT no longer exists.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/google:
    post:
      tags: [Auth]
      summary: Sign in with Google
      description: |-
        Validates a Google ID token and either logs the user in or returns additional profile
        information that the client can use to complete sign-up.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleAuthRequest'
      responses:
        '200':
          description: Google identity was validated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoogleAuthNewUserResponse'
            text/plain:
              schema:
                $ref: '#/components/schemas/JwtToken'
        '401':
          description: The provided Google ID token was invalid.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/sign-up-google:
    post:
      tags: [Auth]
      summary: Prepare a new account using Google Sign-In
      description: |
        Validates a Google ID token for a user that is not yet registered in Strnadi and returns
        the profile information required to finish onboarding together with a provisional JWT.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleAuthRequest'
      responses:
        '200':
          description: Google identity token was accepted and onboarding can continue.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoogleAuthSignupResponse'
        '401':
          description: The provided Google ID token was invalid.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: The e-mail address already belongs to an existing Strnadi account.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login-google:
    post:
      tags: [Auth]
      summary: Sign in an existing account using Google Sign-In
      description: |
        Validates a Google ID token for an existing Strnadi user and returns a freshly issued JWT
        when the account is found. Email verification is marked as completed automatically.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleAuthRequest'
      responses:
        '200':
          description: Google identity was validated and the caller received a JWT.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/JwtToken'
        '401':
          description: The provided Google ID token was invalid.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: No Strnadi account is associated with the Google identity.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/apple:
    post:
      tags: [Auth]
      summary: Sign in or link an Apple ID
      description: |
        Handles Apple sign-in flows for mobile and web clients. When the caller is already
        authenticated the Apple ID will be linked to that account. Otherwise the endpoint issues
        a JWT for onboarding or for an existing user.
      security:
        - {}
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppleAuthRequest'
      responses:
        '200':
          description: Apple identity was validated. When linking an account the response body is empty.
          content:
            application/json:
              schema:
                nullable: true
                oneOf:
                  - $ref: '#/components/schemas/AppleAuthJwtResponse'
                  - $ref: '#/components/schemas/AppleAuthSignupResponse'
        '400':
          description: The request was missing required Apple identity fields or the email could not be verified.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: The provided Apple credentials were invalid or the caller supplied an invalid JWT while linking accounts.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/apple-callback:
    post:
      tags: [Auth]
      summary: Handle the Apple sign-in redirect callback
      description: |
        Accepts the callback that Apple sends after the user authorizes the Strnadi application.
        The endpoint redirects the browser back to the originally requested return URL while
        preserving the `user` payload and the `id_token` issued by Apple in the URI fragment.
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [state]
              properties:
                user:
                  type: string
                  description: Base64 encoded JSON payload with the Apple user information.
                state:
                  type: string
                  description: Encoded state value that contains the original return URL.
                id_token:
                  type: string
                  description: The Apple identity token that should be persisted by the client.
      responses:
        '302':
          description: Redirected to the return URL with the Apple payload appended as a fragment.
        '400':
          description: The callback parameters were invalid or missing.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/apple/callback:
    post:
      tags: [Auth]
      summary: Handle the Apple sign-in redirect for Android clients
      description: |
        Mirrors the web callback but returns an Android `intent://` deep link that forwards the
        Apple payload back into the mobile application.
      security: []
      requestBody:
        required: false
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                user:
                  type: string
                  description: Base64 encoded JSON payload with the Apple user information.
                  nullable: true
                state:
                  type: string
                  description: Encoded state value that contains the original return URL.
                  nullable: true
                id_token:
                  type: string
                  description: Apple identity token supplied by the platform.
                  nullable: true
                code:
                  type: string
                  description: Authorization code provided by Apple when `response_mode=form_post`.
                  nullable: true
      responses:
        '302':
          description: Redirected to the Android intent deep link containing the Apple payload.
  /auth/has-apple-id:
    get:
      tags: [Auth]
      summary: Check whether a user has a linked Apple ID
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
      security: []
      responses:
        '200':
          description: The specified user has an Apple ID linked to the account.
        '409':
          description: The user has no linked Apple ID or does not exist.
  /auth/has-google-id:
    get:
      tags: [Auth]
      summary: Check whether a user has a linked Google identity
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
      security: []
      responses:
        '200':
          description: The specified user has a Google ID linked to the account.
        '409':
          description: The user has no linked Google identity or does not exist.
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login succeeded and a JWT token was issued.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/JwtToken'
        '401':
          description: The password did not match.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: No user with the supplied email exists.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/sign-up:
    post:
      tags: [Auth]
      summary: Sign up with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpRequest'
      responses:
        '200':
          description: Signup succeeded and a JWT was issued.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/JwtToken'
        '400':
          description: The request did not contain the required data or JWT when attempting an Apple linked signup.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: The email address is already registered or the user could not be created.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/{userId}/resend-verify-email:
    get:
      tags: [Auth]
      summary: Resend the email verification link
      parameters:
        - name: userId
          in: path
          required: true
          description: Numeric identifier of the user requesting another verification email.
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: A new verification email has been queued.
        '208':
          description: The email address is already verified.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: The bearer token or the associated email was missing.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: The supplied JWT was invalid or did not belong to the target user.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/{email}/reset-password:
    get:
      tags: [Auth]
      summary: Initiate a password reset email
      description: Sends a password reset email containing a JWT token.
      parameters:
        - name: email
          in: path
          required: true
          description: Email address that should receive the reset message.
          schema:
            type: string
            format: email
      security: []
      responses:
        '200':
          description: A password reset email has been queued.
        '404':
          description: No user with the supplied email exists.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: string
      description: Human readable message describing the error.
    JwtToken:
      type: string
      description: Encoded JSON Web Token.
      example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
    GoogleAuthRequest:
      type: object
      required: [idToken]
      properties:
        idToken:
          type: string
          description: Google ID token obtained from the client SDK.
    GoogleAuthNewUserResponse:
      type: object
      required: [jwt, firstName, lastName, googleId]
      properties:
        jwt:
          $ref: '#/components/schemas/JwtToken'
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        googleId:
          type: string
          description: Stable Google subject identifier returned in the ID token.
    GoogleAuthSignupResponse:
      type: object
      required: [jwt, firstName, lastName]
      properties:
        jwt:
          $ref: '#/components/schemas/JwtToken'
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
    AppleAuthRequest:
      type: object
      required: [idToken, userIdentifier]
      properties:
        idToken:
          type: string
          description: Apple ID token returned by the native or web flow.
        email:
          type: string
          format: email
          nullable: true
        givenName:
          type: string
          nullable: true
        familyName:
          type: string
          nullable: true
        userIdentifier:
          type: string
          description: Stable identifier assigned by Apple to the user.
    AppleAuthJwtResponse:
      type: object
      required: [jwt]
      properties:
        jwt:
          $ref: '#/components/schemas/JwtToken'
    AppleAuthSignupResponse:
      type: object
      required: [jwt, exists]
      properties:
        jwt:
          $ref: '#/components/schemas/JwtToken'
        exists:
          type: boolean
          description: Indicates whether the user already existed in Strnadi before this login.
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        appleid:
          type: string
          description: Unique Apple identifier returned in the ID token claims.
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
...